basic working version complete (w/ some weird padding behavior). Cleanup of code structure and a11y audit needed
